+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Changelog
 Changes in 0.6.X
 ================
 
+- Added Binary support to UUID (MongoEngine/mongoengine#47)
 - Fixed MapField lookup for fields without declared lookups (MongoEngine/mongoengine#46)
 - Fixed BinaryField python value issue (MongoEngine/mongoengine#48)
 - Fixed SequenceField non numeric value lookup (MongoEngine/mongoengine#41)